Handle Windows 7 ipconfig output in French locale
... which could also work for other romance languages, since the major
change here was swapping the order of the adjective and noun in the
'adapater' regex (French uses effectively 'Adapater Ethernet' rather
than 'Ethernet Adapter' which is how a germanic language would order
things).

The `ipconfig` output in French also adds an extra trailing space to
the interface name which needs to be removed so that the hash key for
the ip info hash doesn't contain unnecessary characters.