Work around ip() failures on non-English systems
When returning the IP address of the default interface, the code was
assuming a default key called "Local Area Connection".  This of course
doesn't work on non-English systems.  An improvement upon the original
implementation is simply to return the first item of the sorted list
of interfaces found on the system.  Although this also probably
doesn't solve the underlying issue about which interface to choose
by default, it is certainly an improvement and allows e.g. cygwin
systems running in a non-English locale to now pass the base.t tests.