Remove dead, incorrect code in if_info()
The if_info() should return a hash containing interface names and ip
addresses in key/value pairs.  Theoretically, if if_info() was called as
a function, then it would have returned the string of the `ifconfig`
binary, which is incorrect.  The functional version of if_info() isn't
exported, hence it couldn't be used anyway, but further than that, if it
had been used via its functional interface, it would have thrown an
error, since $self would be a reference and hence the functional code
would have barfed.  An inspection of the code showed that this code path
was never taken, hence the offending code could simply be removed.