Name base test files more correctly
Actually, the base tests are simply "base" for that system, so they
aren't the base tests for unix.  The base tests for Windows are really
the mocked windows tests, so they should be described as such.  Hence,
when testing on a Windows system, the base tests will test the actual
Windows system and the mocked tests will run for fake Windows systems.