Fall back to 'ip' if 'ifconfig' isn't available
On linux systems the `ifconfig` command is deprecated and some systems
don't have this command anymore.  In such situations, we should fall
back to using the `ip` command, which in that case *should* exist.

In order to implement this nicely, I used ideas already submitted by
@AdamBalali in pull request #5 and added a test which uses a mock linux
`ip` command output to fake a system where only the `ip` command exists.