Fix failures on Russian Windows XP systems
There was a failure on CPANTESTERS
(https://www.cpantesters.org/cpan/report/7b9efe16-6bff-1014-9cac-c182dcff62bf)
which was due to the regex which checks for the interface info on
Windows not matching correctly.  This then returned an empty list of
keys for the `%if_info` hash and hence caused the code to break.  After
the user (chorny++) gave me the output which caused the problem, it was
possible to work out what the problem was and create a solution.  I also
converted his original file into UTF8 via `iconv`

```
iconv -t UTF8 -f cp1251 ipconfig_output_ru > ipconfig-winxp-ru-utf8.txt
```

in order that the text could be read in a UTF8 locale and this file has
been added to the test suite.